% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BarkFunction}
\alias{BarkFunction}
\title{BarkFunction}
\usage{
BarkFunction(
  DiameterValue,
  SpeciesGroupKey,
  SpeciesGroupDefinition,
  Top_ob,
  DBH,
  LogLength
)
}
\arguments{
\item{DiameterValue}{numeric vector of corresponding diameters (mm)}

\item{SpeciesGroupKey}{Species ID}

\item{SpeciesGroupDefinition}{List of species group information, with speciesgroupkey as the name of the elements(see getSpeciesGroupDefinition)}

\item{Top_ob}{Starting position of log along the stem}

\item{DBH}{in mm, for  Skogforsk 2004 barkFunction categories}

\item{LogLength}{Optional, in cm}
}
\value{
Log volume in m3
}
\description{
Calculates diametervalues under bark
}
\seealso{
Buck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
