% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckHpr.R
\name{buckHpr}
\alias{buckHpr}
\title{buckHpr}
\usage{
buckHpr(
  XMLNode,
  PriceMatrices,
  ProductData,
  StemProfile,
  PermittedGrades,
  SpeciesGroupDefinition
)
}
\arguments{
\item{XMLNode}{ouput from getXMLNode()}

\item{PriceMatrices}{list of price matrices for all ProductKeys (getPriceMatrices())}

\item{ProductData}{Matrix containing product data (getProductData())}

\item{StemProfile}{Stem profiles for all stems in hprfile (getStemProfile())}

\item{PermittedGrades}{list with the same lenght of assortments, each element containing the stemgrades allowed in each assortment (getPermittedGrades())}

\item{SpeciesGroupDefinition}{getSpeciesGroupDefinition()}
}
\value{
result structure with optimum bucking solution for the stems in the input .hpr file
}
\description{
Calculate optimal bucking for all stems in a hpr file
}
\references{
Skogforsk 2011. Introduction to StanForD 2010. URL: Skogforsk. https://www.skogforsk.se/contentassets/1a68cdce4af1462ead048b7a5ef1cc06/stanford-2010-introduction-150826.pdf
}
\seealso{
getPermittedGrades, getPriceMatrices, getProductData
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
