% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{equal.lengths}
\alias{equal.lengths}
\title{equal.lengths}
\usage{
equal.lengths()
}
\value{
error when vector lengths are not equal
}
\description{
Test whether vectors are of equal lengths
}
\examples{
a=c(1,2);b=c(1,2,3)
equal.lengths(a,b)
}
\seealso{
OptApt
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
