% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBucking}
\alias{getBucking}
\title{getBucking}
\usage{
getBucking(hprfile, PriceMatrices, ProductData, StemProfile)
}
\arguments{
\item{hprfile}{Path to .hpr file}

\item{PriceMatrices}{list of prices matrices for all ProductKeys (see getPriceMatrices())}

\item{ProductData}{Matrix containing product data (see getProductData())}

\item{StemProfile}{Stem profiles for all stems in hprfile (see getStemProfile())}
}
\value{
Output structure with bucking outcomes
}
\description{
Extract bucking outcomes from a .hpr file
}
\seealso{
OptBuck, Optbuck_hpr
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
