% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getHarvestedArea}
\alias{getHarvestedArea}
\title{getHarvestedArea}
\usage{
getHarvestedArea(Stems)
}
\arguments{
\item{Stems}{output of getStems()}
}
\value{
Simple feature object of area around harvested trees
}
\description{
Extract harvested area
}
\examples{
Stems=getStems(hprfile)
getHarvestedArea(Stems)
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
