% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLengthClasses.R
\name{getLengthClasses}
\alias{getLengthClasses}
\title{getLengthClasses}
\usage{
getLengthClasses(XMLNode)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
List of length classes for assortments, element names correspond to product keys
}
\description{
Extract the length classes for each assortment from .hpr files, needed for volume calculation when VolumeLengthCategory=="Length as defined in LengthClasses"
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
