% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getLogs}
\alias{getLogs}
\title{getLogs}
\usage{
getLogs(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
data table with log information
}
\description{
Extract information on harvested logs from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
