% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPermittedGrades}
\alias{getPermittedGrades}
\title{getPermittedGrades}
\usage{
getPermittedGrades(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
List of permitted grades for assortments, element names correspond to product keys
}
\description{
Extract the permitted stem grades for each assortment from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
