% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPriceMatrices.R
\name{getPriceMatrices}
\alias{getPriceMatrices}
\title{getPriceMatrices}
\usage{
getPriceMatrices(XMLNode)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
list of prices matrices for all ProductKeys. Element names are productkeys.
}
\description{
Extract product data from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
