% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProductData.R
\name{getProductData}
\alias{getProductData}
\title{getProductData}
\usage{
getProductData(XMLNode, SpeciesGroupDefinition)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
Information on ProductKeys, ProductNames, ProductGroupName, SpeciesGroupKey, DiameterUnderBark, DiameterClassLowerLimit, DiameterClassMAX, LengthClassLowerLimit, LengthClassMAX, VolumeDiameterCategory, DiameterTopPositions
}
\description{
Extract product data from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
