% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSortimentOverview}
\alias{getSortimentOverview}
\title{getSortimentOverview}
\usage{
getSortimentOverview(Logs, ProductData)
}
\arguments{
\item{Logs}{otput from getLogs}

\item{ProductData}{output from getProductData}
}
\value{
figure in viewer
}
\description{
Extract distribution of harvested volume over assortments
}
\seealso{
getLogs, getProductData
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
