% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeciesGroupDefinition.R
\name{getSpeciesGroupDefinition}
\alias{getSpeciesGroupDefinition}
\title{getSpeciesGroupDefinition}
\usage{
getSpeciesGroupDefinition(XMLNode)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
List of species group information, with speciesgroupkey as the name of the elements
}
\description{
Extract information on species groups from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
