% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStemprofile.R
\name{getStemprofile}
\alias{getStemprofile}
\title{getStemprofile}
\usage{
getStemprofile(XMLNode, Logs)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}

\item{Logs}{Harvested logs (getLogs())}
}
\value{
Stem profiles of harvested stems with stem grades
}
\description{
Extract stem profiles from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
