% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStems.R
\name{getStems}
\alias{getStems}
\title{getStems}
\usage{
getStems(XMLNode)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
data table with stem information
}
\description{
Extract information on harvested stems from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
