% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMLNode.R
\name{getXMLNode}
\alias{getXMLNode}
\title{getXMLNode}
\usage{
getXMLNode(hprfile)
}
\arguments{
\item{hprfile}{Path to hpr file}
}
\value{
object of class XMLNode, parsed from the hpr file
}
\description{
Parse the hpr file
}
\references{
https://cran.r-project.org/web/packages/XML/XML.pdf
}
