% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.whole}
\alias{is.whole}
\title{is.whole}
\usage{
is.whole(a, tol = 1e-07)
}
\arguments{
\item{a}{A number}

\item{tol}{Tolerance}
}
\value{
Logical: "True" if whole and "False" if decimal
}
\description{
Test if number(s) is/are whole or decimal
}
\examples{
a=c(1,2);b=1.2
is.whole(a)
is.whole(b)
}
\seealso{
OptApt
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
