% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optBuck_hpr}
\alias{optBuck_hpr}
\title{optBuck_hpr}
\usage{
optBuck_hpr(
  hprfile,
  PriceMatrices,
  ProductData,
  StemProfile,
  PermittedGrades,
  ...
)
}
\arguments{
\item{hprfile}{Path to input .hpr file}

\item{PriceMatrices}{list of price matrices for all ProductKeys (see getPriceMatrices)}

\item{ProductData}{Matrix containing product data (see getProductData)}

\item{StemProfile}{Stem profiles for all stems in hprfile (see getStemProfile)}

\item{PermittedGrades}{list with the same lenght of assortments, each element containing the stemgrades allowed in each assortment (see getPermittedGrades)}

\item{...}{others}
}
\value{
result structure with optimum bucking solution for the stems in the input hpr file
}
\description{
Calculate optimal bucking for hpr files
}
\references{
Skogforsk 2011. Introduction to StanForD 2010. URL: Skogforsk. https://www.skogforsk.se/contentassets/1a68cdce4af1462ead048b7a5ef1cc06/stanford-2010-introduction-150826.pdf
}
\seealso{
getPermittedGrades, getPriceMatrices, getProductData
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
