% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotBucking}
\alias{plotBucking}
\title{plotBucking}
\usage{
plotBucking(Bucking, StemProfile, Key)
}
\arguments{
\item{Bucking}{output structure of getBucking(), optBuck() or optBuck_hpr()}

\item{StemProfile}{StemProfile (see getStemprofile())}

\item{Key}{StemKey of the stem to be plotted}
}
\value{
plot of bucking outcome
}
\description{
Plot the bucking outcome
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
