% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictStemprofile.R
\name{predictStemprofile}
\alias{predictStemprofile}
\title{predictStemprofile}
\usage{
predictStemprofile(
  XMLNode,
  ProductData,
  PermittedGrades,
  GradeDetermination = "GradeValue"
)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}

\item{ProductData}{output of getProductData()}

\item{PermittedGrades}{output of getPermittedGrades()}

\item{GradeDetermination}{Either "GradeValue" or "LogsProductKey". Shoud stem grade be based on StemGrade and gradeStartPosition or selected Products of logs}
}
\value{
Output structure with stem profile containing stem grades
}
\description{
Predict and extract Norway spruce stem profiles using taper models based on the log dimensions, for cases when no stem profile is recorded in the hpr file.
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
