% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PriceVolumeCalc}
\alias{PriceVolumeCalc}
\title{PriceVolumeCalc}
\usage{
PriceVolumeCalc(
  VolumeDiameterAdjustment,
  VolumeDiameterCategory,
  VolumeLengthCategory,
  diameterPosition,
  DiameterValue,
  StartPos,
  StopPos,
  DiameterTopPosition,
  DiameterUnderBark = T,
  SpeciesGroupKey = NA,
  SpeciesGroupDefinition = NA,
  DBH = NA,
  LogLength = NA,
  LengthClasses = NA,
  ProductKey = NA
)
}
\arguments{
\item{VolumeDiameterAdjustment}{Volume diameter adjustment according to stanford2010 (see getProductData()).}

\item{VolumeDiameterCategory}{Volume calculation method according to stanford2010 (see getProductData()).}

\item{VolumeLengthCategory}{Volume length category according to stanford2010 (see getProductData()).}

\item{diameterPosition}{numeric vector of diameter positions (cm) of a stem profile; 0,10,...,end}

\item{DiameterValue}{numeric vector of corresponding diameters (mm)}

\item{StartPos}{Starting position of log along the stem}

\item{StopPos}{Ending position of log}

\item{DiameterTopPosition}{Position from top end of log where top diameter is measured. Cm}

\item{DiameterUnderBark}{Logical TRUE/FALSE}

\item{SpeciesGroupKey}{Species ID}

\item{SpeciesGroupDefinition}{List of species group information, with speciesgroupkey as the name of the elements(see getSpeciesGroupDefinition)}

\item{DBH}{Optional, in mm (see BarkFunction)}

\item{LogLength}{Optional, in cm (see BarkFunction)}

\item{LengthClasses}{List of length classes for the assortments (see getLengthClasses)}

\item{ProductKey}{Assortment key (see getProductData())}
}
\value{
Log volume in m3
}
\description{
Calculates log price volume, i.e., the volume which is used for price calculation
}
\seealso{
Buck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
