% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StemprofileIncrement}
\alias{StemprofileIncrement}
\title{StemprofileIncrement}
\usage{
StemprofileIncrement(Stemprofile, DBH2, breastheight)
}
\arguments{
\item{Stemprofile}{Stem profiles for all stems in hprfile (see getStemProfile)}

\item{DBH2}{a numeric vector of new DBHs, of the same length as unique StemKeys in Stemprofile}

\item{breastheight}{height in cm which is considered breastheight (numeric), typically 110 or 130.}
}
\value{
A new Stemprofile object in which the new diameters are added
}
\description{
Predict Stemprofile at another point in time given a vector of new DBHs
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
