% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strsplits}
\alias{strsplits}
\title{strsplits}
\usage{
strsplits(x, splits)
}
\arguments{
\item{x}{character vector to split}

\item{splits}{vector of character patterns used to split}
}
\value{
List of permitted grades for assortments
}
\description{
modified strsplit for multiple splits
}
\seealso{
getPriceMatrices
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
