% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{track_trace}
\alias{track_trace}
\title{track_trace}
\usage{
track_trace(m, tt)
}
\arguments{
\item{m}{matrix of potential cuts}

\item{tt}{matrix of log segment which maximize cumulative value}
}
\value{
Logical: "True" if whole and "False" if decimal
}
\description{
Back-track optimum bucking solution
}
\seealso{
OptApt
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
