% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VolumeCalc}
\alias{VolumeCalc}
\title{VolumeCalc}
\usage{
VolumeCalc(
  diameterPosition,
  DiameterValue,
  StartPos,
  StopPos,
  DiameterTopPosition,
  DiameterUnderBark = T,
  SpeciesGroupKey = NA,
  SpeciesGroupDefinition = NA,
  DBH = NA,
  LogLength = NA
)
}
\arguments{
\item{diameterPosition}{numeric vector of diameter positions (cm) of a stem profile; 0,10,...,end}

\item{DiameterValue}{numeric vector of corresponding diameters (mm)}

\item{StartPos}{Starting position of log along the stem}

\item{StopPos}{Ending position of log}

\item{DiameterTopPosition}{Position from top end of log where top diameter is measured. Cm}

\item{DiameterUnderBark}{Logical TRUE/FALSE}

\item{SpeciesGroupKey}{Species ID}

\item{SpeciesGroupDefinition}{List of species group information, with speciesgroupkey as the name of the elements(see getSpeciesGroupDefinition)}

\item{DBH}{Optional, in mm (see BarkFunction)}

\item{LogLength}{Optional, in cm (see BarkFunction)}
}
\value{
Log volume in m3
}
\description{
Calculates log volume from all diameters as solid volume
}
\seealso{
Buck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
