% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarkFunction.R
\name{BarkFunction}
\alias{BarkFunction}
\title{BarkFunction}
\usage{
BarkFunction(
  DiameterValue,
  SpeciesGroupKey,
  SpeciesGroupDefinition,
  Top_ob,
  DBH,
  LogLength
)
}
\arguments{
\item{DiameterValue}{Numeric vector of diameters (mm)}

\item{SpeciesGroupKey}{Species ID (getProductData())}

\item{SpeciesGroupDefinition}{List of species group information, with speciesgroupkey as the name of the elements (getSpeciesGroupDefinition())}

\item{Top_ob}{Top diameter over bark}

\item{DBH}{in mm, for  Skogforsk 2004 barkFunction categories}

\item{LogLength}{in cm}
}
\value{
Log volume in m3
}
\description{
Calculates diametervalues under bark
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
