% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBucking.R
\name{getBucking}
\alias{getBucking}
\title{getBucking}
\usage{
getBucking(XMLNode, PriceMatrices, ProductData, StemProfile, LengthClasses)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}

\item{PriceMatrices}{list of prices matrices for all ProductKeys (getPriceMatrices())}

\item{ProductData}{Matrix containing product data (getProductData())}

\item{StemProfile}{Stem profiles for all stems in hprfile (getStemProfile())}

\item{LengthClasses}{getLengthClasses()}
}
\value{
Output structure with bucking outcomes
}
\description{
Extract bucking outcomes from a .hpr file
}
\seealso{
buckStem, buckHpr
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
