% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getLengthClasses}
\alias{getLengthClasses}
\title{LengthClasses}
\usage{
getLengthClasses(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
List of length classes for assortments, element names correspond to product keys
}
\description{
Extract the length classes for each assortment from .hpr files, needed for volume calculation when VolumeLengthCategory=="Length as defined in LengthClasses"
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
