% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogs.R
\name{getLogs}
\alias{getLogs}
\title{getLogs}
\usage{
getLogs(XMLNode)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
data table with log information
}
\description{
Extract information on harvested logs from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
