% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPermittedGrades.R
\name{getPermittedGrades}
\alias{getPermittedGrades}
\title{getPermittedGrades}
\usage{
getPermittedGrades(XMLNode)
}
\arguments{
\item{XMLNode}{Output of getXMLNode()}
}
\value{
List of permitted grades for assortments, element names correspond to product keys
}
\description{
Extract the permitted stem grades for each assortment from .hpr files
}
\seealso{
buckStem
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
