% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPriceMatrices}
\alias{getPriceMatrices}
\title{getPriceMatrices}
\usage{
getPriceMatrices(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
list of prices matrices for all ProductKeys. Element names are productkeys.
}
\description{
Extract product data from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
