% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getProductData}
\alias{getProductData}
\title{getProductData}
\usage{
getProductData(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
Information on ProductKeys, ProductNames, ProductGroupName, SpeciesGroupKey, DiameterUnderBark, DiameterClassLowerLimit, DiameterClassMAX, LengthClassLowerLimit, LengthClassMAX, VolumeDiameterCategory, DiameterTopPositions
}
\description{
Extract product data from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
