% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSpeciesGroupDefinition}
\alias{getSpeciesGroupDefinition}
\title{getSpeciesGroupDefinition}
\usage{
getSpeciesGroupDefinition(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
List of species group information, with speciesgroupkey as the name of the elements
}
\description{
Extract information on species groups from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
