% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getStemprofile}
\alias{getStemprofile}
\title{getStemprofile}
\usage{
getStemprofile(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}

\item{Logs}{Harvested logs (getLogs())}
}
\value{
Stem profiles of harvested stems with stem grades
}
\description{
Extract stem profiles from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
