% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getStems}
\alias{getStems}
\title{getStems}
\usage{
getStems(hprfile)
}
\arguments{
\item{hprfile}{Path to input .hpr file}
}
\value{
data table with stem information
}
\description{
Extract information on harvested stems from .hpr files
}
\seealso{
optBuck
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
