% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotBucking}
\alias{plotBucking}
\title{plotBucking}
\usage{
plotBucking(diameterPosition, DiameterValue, StemGrade, res)
}
\arguments{
\item{diameterPosition}{vector of diameter positions (cm) of a stem profile: 0,10,...,end}

\item{DiameterValue}{vector of corresponding diameters (mm) for those diameter positions}

\item{StemGrade}{vector of corresponding stem grades}

\item{res}{the bucing outcome, i.e., output from optBuck()}
}
\value{
plot of bucking outcome
}
\description{
Plot the bucking outcome
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
