% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBucking.R
\name{plotBucking}
\alias{plotBucking}
\title{plotBucking}
\usage{
plotBucking(Res, StemProfile, Stem, ProductData)
}
\arguments{
\item{StemProfile}{StemProfile (getStemprofile())}

\item{Bucking}{output structure of getBucking(), buckStem() or buckHpr()}

\item{Key}{StemKey of the stem to be plotted}
}
\value{
plot of bucking outcome
}
\description{
Plot the bucking outcome
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
