% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predictStemprofile}
\alias{predictStemprofile}
\title{predictStemprofile}
\usage{
predictStemprofile(hprfile, ProductData, PermittedGrades)
}
\arguments{
\item{hprfile}{Path to .hpr file}

\item{ProductData}{output of getProductData()}

\item{PermittedGrades}{output of getPermittedGrades()}
}
\value{
Output structure with stem profile containing stem grades
}
\description{
Predict and extract Norway spruce stem profiles using taper models based on the log dimensions, for cases when no stem profile is recorded in the hpr file.
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
