% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarkFunction.R
\name{strSplits}
\alias{strSplits}
\title{strSplits}
\usage{
strSplits(x, splits)
}
\arguments{
\item{x}{character vector to split}

\item{splits}{vector of character patterns used to split}
}
\value{
List of permitted grades for assortments
}
\description{
Helper function: modified strsplit() for multiple splits
}
\seealso{
getPriceMatrices
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
